EncodingTest movido para CertificateBuilderTest. Classes padronizadas.
